refactor: class LottoDao, class LottoGameDao
singleton 패턴 적용
